db test added

+++ core-logic/src/main/java/lk/dinuka/core/impl/UserRepositoryImpl.java
@@ -18,7 +18,7 @@ public class UserRepositoryImpl implements UserRepository {
     @Override
     public void saveUser(DomainUser domainUser) {
         String query = "INSERT INTO user (userId, memberType, userName, userEmail) " +
-                "VALUES ('"+ domainUser.getUserId() + "','" + domainUser.getMemberType() + "','" +
+                "VALUES ('" + domainUser.getUserId() + "','" + domainUser.getMemberType() + "','" +
                 domainUser.getUserName() + "','" + domainUser.getUserEmail() + "');";
 
         jdbcTemplate.update(query);
@@ -35,20 +35,31 @@ public class UserRepositoryImpl implements UserRepository {
 
     @Override
     public DomainUser getUserById(int userId) {
-        String sql = "SELECT * FROM user WHERE userId = ?";
-        DomainUser user = null;
+        String sql = "SELECT * FROM user WHERE userId = " + userId;
+        DomainUser user = jdbcTemplate.query(sql, new DomainUserMapper()).get(0);
         return user;
     }
 
     @Override
     public void deleteUserById(int userId) {
-        String sql = "DELETE FROM user WHERE userId = ?";
-
+        String sql = "DELETE FROM user WHERE userId = "+userId;
+        jdbcTemplate.execute(sql);
     }
 
     @Override
     public void deleteAllUsers() {
+        String sql = "truncate table user;";
+        jdbcTemplate.execute(sql);
+    }
 
+    private static final class DomainUserMapper implements RowMapper<DomainUser> {
+
+        public DomainUser mapRow(ResultSet rs, int rowNum) throws SQLException {
+            DomainUser user = new DomainUser(rs.getInt("userId"), rs.getString("userName"),
+                    rs.getString("userEmail"),
+                    rs.getString("memberType"));
+            return user;
+        }
     }
 
     public JdbcTemplate getJdbcTemplate() {
@@ -57,14 +68,6 @@ public class UserRepositoryImpl implements UserRepository {
 
     public void setJdbcTemplate(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
-    }
-
-    private static final class DomainUserMapper implements RowMapper<DomainUser> {
-
-        public DomainUser mapRow(ResultSet rs, int rowNum) throws SQLException {
-            DomainUser user = new DomainUser(rs.getInt("userId"), rs.getString("userName"),rs.getString("userEmail"),rs.getString("memberType"));
-            return user;
-        }
     }
 
 }
